split ui and view components
