+++ packages/frontend/pages/index.js
@@ -7,6 +7,7 @@ function Index() {
     <div>
       <Head>
         <title>Uçurtma</title>
+        {/* TODO: should we use google fonts? there is advantages and disadvantages of using google fonts but i don't think we have to. */}
         <link
           href="https://fonts.googleapis.com/css?family=Quicksand:400,500,700&display=swap&subset=latin-ext"
           rel="stylesheet"
